More bugs

+++ server.py
@@ -35,7 +35,7 @@ def index():
             if fn in repo.ignored_files: continue
             repo.branches[fn.name] = nightlies.Branch(repo, fn.name)
 
-    if "pid" in current_process:
+    if current_process and "pid" in current_process:
         try:
             # Does not actually kill, but does check if pid exists
             os.kill(current_process["pid"], 0)
